[Finish]応用課題3

+++ app/controllers/book_comments_controller.rb
@@ -17,6 +17,9 @@ class BookCommentsController < ApplicationController
 	def destroy
 		book = Book.find(params[:book_id])
 		comment = current_user.book_comments.find_by(book_id: book.id)
+		if current_user.id != comment.user_id
+			redirect_to book_path(book)
+		end
 		comment.destroy
         redirect_back(fallback_location: root_path)
 	end
